requirements.txt
transformers
ipykernel
gradio
numpy
python-dotenv
torch